refactor: 7/9 코드리뷰 반영 (인터셉터 의존성 수정)
