+++ src/main/java/gift/interceptor/MemberInterceptor.java
@@ -1,6 +1,6 @@
 package gift.interceptor;
 
-import gift.service.MemberService;
+import gift.util.TokenValidator;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import org.springframework.lang.NonNull;
@@ -10,15 +10,15 @@ import org.springframework.web.servlet.HandlerInterceptor;
 @Component
 public class MemberInterceptor implements HandlerInterceptor {
 
-    private final MemberService memberService;
+    private final TokenValidator tokenValidator;
 
-    public MemberInterceptor(MemberService memberService) {
-        this.memberService = memberService;
+    public MemberInterceptor(TokenValidator tokenValidator) {
+        this.tokenValidator = tokenValidator;
     }
 
     @Override
     public boolean preHandle(@NonNull HttpServletRequest request, @NonNull HttpServletResponse response, @NonNull Object handler) throws Exception {
-        memberService.validateToken(request);
+        tokenValidator.validateToken(request);
         return true;
     }
 }
